requirements.txt
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.6.0
pymongo==3.11.0
python-decouple==3.3
python-dotenv==0.14.0
toml==0.10.1
Werkzeug==1.0.1
